api/requirements.txt
fastapi
pyyaml
torch
torchaudio
numpy 